backend/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.6.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
sqlparse==0.5.3
celery
redis
psycopg2-binary
google-api-python-client
python-dotenv